Unnötige Methoden entfernt + Logout funktion erstellt
unnötige Methoden entfernt
logout funktion erstellt
im register controller das /menu weggenommen das brauchen wir nicht

success jsp abgeändert kann aber eig. gelöscht werden komplett

successlogin jsp gelöscht
