+++ WebContent/WEB-INF/jsp/sidemenu.jsp
@@ -12,10 +12,10 @@
 <body>
 	<div id='cssmenu'>
 		<ul>
-			<li><a href="<c:url value='/menu/home' />"><span><img src="<c:url value="static/images/home-3-48.png"/>"/></span></a></li>
-			<li><a href="<c:url value='/list' />"><span><img src="<c:url value="static/images/list-2-48.png"/>"/></span></a>
+			<li><a href="<c:url value='/home' />"><span><img src="<c:url value="static/images/home-3-48.png"/>"/></span></a></li>
+			<li><a href="<c:url value='/userlist' />"><span><img src="<c:url value="static/images/list-2-48.png"/>"/></span></a>
 			<li class='last'><a href="<c:url value='/login' />"><span><img src="<c:url value="static/images/login-48.png"/>"/></span></a></li>
-			<li><a href="<c:url value='/menu/logout' />"><span><img src="<c:url value="static/images/logout-48.png"/>"/></span></a></li>			
+			<li><a href="<c:url value='/logout' />"><span><img src="<c:url value="static/images/logout-48.png"/>"/></span></a></li>			
 		</ul>
 		
 		<ul>
