+++ src/de/szut/ProjectZer0/web/LoginController.java
@@ -50,21 +50,7 @@ public class LoginController {
 		} else {
 			req.getSession().setAttribute("user", user);
 			// send 'successful login' screen
-			return "redirect:ListOfUser";
-		}
-	}
-
-	@RequestMapping("/ListOfUser")
-	private String returnUserList(HttpServletRequest req) {
-		User user = (User) req.getSession().getAttribute("user");
-		System.out.println(user.getUsername());
-		if(req.getSession().getAttribute("user") != null)
-		{
-			return "userList";
-		}
-		else
-		{
-			return "error";
+			return "redirect:home";
 		}
 	}
 
@@ -76,18 +62,6 @@ public class LoginController {
 		return null;
 	}
 
-	@RequestMapping("/loginError")
-	public ModelAndView erpLoginError() {
-		ModelMap map = new ModelMap();
-		return new ModelAndView("error", map);
-	}
-
-	@RequestMapping("/successLogin")
-	public ModelAndView erpLoginSuccessfull() {
-		ModelMap map = new ModelMap();
-		return new ModelAndView("successLogin", map);
-	}
-
 	// Testdaten hinzufügen TODO: Mehrere Datenpakete
 	@RequestMapping("/generateTestData")
 	public ModelAndView generateData() {
@@ -97,14 +71,18 @@ public class LoginController {
 
 		return new ModelAndView("redirect:login", map);
 	}
-
-	/*static boolean validateUser(HttpServletRequest req) {
-		Object userAttribute = req.getSession().getAttribute("user");
-		if (user == userAttribute) {
-			return true;
-		} else {
-			return false;
+	
+	
+	@RequestMapping("/logout")
+	private String returnUserList(HttpServletRequest req) {
+		if(req.getSession().getAttribute("user") != null)
+		{
+			req.getSession().setAttribute("user", null);
+			return "login";
+		}
+		else
+		{
+			return "error";
 		}
 	}
-	*/
 }
